+++ packages/contracts/contracts/interfaces/ISpend2Verifier.sol
@@ -1,0 +1,13 @@
+//SPDX-License-Identifier: GPL-3.0
+pragma solidity 0.7.6;
+
+/// @title Verifier interface.
+/// @dev Interface of Verifier contract.
+interface ISpend2Verifier {
+    function verifyProof(
+        uint256[2] memory a,
+        uint256[2][2] memory b,
+        uint256[2] memory c,
+        uint256[9] memory input
+    ) external view returns (bool);
+}