+++ packages/sdk/src/proof/solidity.ts
@@ -1,0 +1,25 @@
+import { BaseProof } from "./common";
+
+export type SolidityProof = [
+  bigint,
+  bigint,
+  bigint,
+  bigint,
+  bigint,
+  bigint,
+  bigint,
+  bigint
+];
+
+export default function packToSolidityProof(proof: BaseProof): SolidityProof {
+  return [
+    proof.pi_a[0],
+    proof.pi_a[1],
+    proof.pi_b[0][1],
+    proof.pi_b[0][0],
+    proof.pi_b[1][1],
+    proof.pi_b[1][0],
+    proof.pi_c[0],
+    proof.pi_c[1],
+  ];
+}