+++ packages/contracts/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ### Unreleased
 
+- Add tests for verifier contract
 - Make commitment tree and hash functions generic behind interfaces
 - Add contracts as package in yarn workspace